refactor(server): migrate to controller style
