+++ server/src/main/kotlin/dev/yidafu/blog/common/controller/CommonControllerModule.kt
@@ -1,8 +1,10 @@
-package dev.yidafu.blog.common.handler
+package dev.yidafu.blog.common.controller
 
 import org.koin.core.annotation.ComponentScan
 import org.koin.core.annotation.Module
 
+
 @Module
 @ComponentScan
-class CommonHandlerModule
+class CommonControllerModule
+
