+++ server/src/main/kotlin/dev/yidafu/blog/fe/FrontendVerticle.kt
@@ -1,7 +1,6 @@
 package dev.yidafu.blog.fe
 
 import dev.yidafu.blog.common.routes.mountPublicRoutes
-import dev.yidafu.blog.fe.routes.mountPostRoutes
 import io.vertx.ext.web.Router
 import io.vertx.ext.web.handler.LoggerHandler
 import io.vertx.kotlin.coroutines.CoroutineRouterSupport
@@ -9,6 +8,8 @@ import io.vertx.kotlin.coroutines.CoroutineVerticle
 import io.vertx.kotlin.coroutines.coAwait
 import org.koin.core.Koin
 import org.slf4j.LoggerFactory
+import dev.yidafu.blog.common.controller.createRoutes as createCommonRouter
+import dev.yidafu.blog.fe.controller.createRoutes as createFeRouter
 
 class FrontendVerticle(private val koin: Koin) : CoroutineVerticle(), CoroutineRouterSupport {
   private val log = LoggerFactory.getLogger(FrontendVerticle::class.java)
@@ -22,7 +23,8 @@ class FrontendVerticle(private val koin: Koin) : CoroutineVerticle(), CoroutineR
       router.route().handler(LoggerHandler.create())
 
       mountPublicRoutes(router)
-      mountPostRoutes(router, koin)
+      createCommonRouter(router)
+      createFeRouter(router)
 
       router.errorHandler(404) { ctx ->
         ctx.end("<h1>404 Not Found</h1>")
