+++ server/src/main/kotlin/dev/yidafu/blog/fe/controller/FeControllerModule.kt
@@ -1,0 +1,6 @@
+package dev.yidafu.blog.fe.controller
+
+import org.koin.core.annotation.Module
+
+@Module
+class FeControllerModule
