--- server/src/main/kotlin/dev/yidafu/blog/fe/handler/FeHandlerModule.kt
@@ -1,8 +1,0 @@
-package dev.yidafu.blog.fe.handler
-
-import org.koin.core.annotation.ComponentScan
-import org.koin.core.annotation.Module
-
-@Module
-@ComponentScan
-class FeHandlerModule
